pyproject.toml
[project]
name = "backend-ai"
version = "0.1.0"
description = "FastAPI Board Service with SQLModel"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "sqlmodel>=0.0.31",
    "pydantic-settings>=2.12.0",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.21",
    "python-jose>=3.5.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "httpx>=0.28.1",
    "ruff>=0.14.10",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

